really awful randomise chord/pattern feature, plays a triad on keys in auto mode and the held keys otherwise... where now?

+++ MidiOutput.hpp
@@ -111,10 +111,9 @@ void midi_bass_send_note_off(int pitch, int velocity, int channel) {
 
 void midi_kill_notes() {
   MIDIOUT.sendControlChange (123, 0, MUSO_GATE_CHANNEL);   // 123 = kill all notes
-  MIDIOUT.sendControlChange (123, 0, MIDI_CHANNEL_BASS_OUT); // 123 = kill all notes
-  MIDIOUT.sendControlChange (123, 0, MIDI_CHANNEL_BITBOX_OUT); // 123 = kill all notes
-  MIDIIN.sendControlChange  (123, 0, MIDI_CHANNEL_BASS_OUT); // 123 = kill all notes - for midiecho to host
-  MIDIIN.sendControlChange  (123, 0, GM_CHANNEL_DRUMS);     // 123 = kill all notes - for midiecho to host
+  if (midiecho_enabled)
+    MIDIIN.sendControlChange (123, 0, GM_CHANNEL_DRUMS);     // 123 = kill all notes - for midiecho to host
+  harmony.kill_notes();
 }
 
 void kill_notes() {
